Lot's of changes throw files, because of trying to change the way of storing data (see description)
The first plan was to implement a simplified version of the model (with calculation errors), and later improve the implementation. But after starting writing code it became clear, that reimplementing will be unreasonably difficult, so now I start writing the final version. Actually we're going to get further from the moment we'll be ready to compile the project and get a result, but in total this will save time (I hope)