Fixed problems in local implementations of atomics
The local implementations of `atomicCAS` were modified some time ago, which brought some errors to it. I've added static asserts about the types used in the implementation and removed type duplication with a `typedef`. The implementation of `atomicAdd` for `double`s is still strange (it has issues with unsafe casts), but it was copied directly from the CUDA documentation, so let it remain as is